requirements.txt
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
open_clip_torch==2.31.0
lancedb==0.20.0
opencv-python==4.11.0.86
pillow==11.1.0
torch==2.6.0
transformers==4.49.0


